# api-spec.yml
openapi: 3.0.0
info:
  title: Kaleidoscope API
  description: >-
    Api for interfacing with the backend of the kaleidoscope project.
    Basically just a few endpoints to pull image data from.
  version: 1.0.0
tags:
  - name: Comments
    description: Comments endpoints

paths:
  /image:
    get:
      summary: Fetch an image by fully qualified file name.
      description: >
        Get the direct file reference to whatever was passed in the query param.
        Returns either the file, or a 404 if an invalid value was passed in.
      parameters:
        - in : query
          name: fileName
          schema:
            type: string
          required: true
          description: Filename including full directory path.
      responses:
        '200':
          description: Successful request, image file fully returned.
        '206':
          description: Partial streaming response, more data incoming.
        '404':
          description: Invalid filename passed, no file found.
  /image-list:
    get:
      summary: Get all images available within the provided directory (optional).
      description: >
        Fetch all images that exist within a given (optional) subdirectory.
        If no directory is provided, it is assumed that you want all images.
      parameters:
        - in : query
          name: subDir
          schema:
            type: string
          required: false
          description: Subdirectory to pull images from.
      responses:
        '200':
          description: 'Returns a string array of fully qualified file names.'
        '404':
          description: 'Invalid directory provided.'
  /available-directories:
    get:
      summary: List of all subdirectories.
      description: >
        Returns a string array of all subdirectories that images could be stored in.
        This generally represents categories of files.
        A root directory path will always be returned along with actual options,
        as a way to select all possible files in other endpoints.
      responses:
        '200':
          description: 'A string array response body.'
  /refresh:
    get:
      summary: Refresh available images
      description: >
        Refreshes the available list of images that the backend will provide.
        Images are maintained in a static lookup list to prevent read operations, so this endpoint will
        trigger a full reload. This should be triggered after images are newly downloaded/added.
      responses:
        '200':
          description: 'Image list was reloaded'